User job page

+++ app/Http/Controllers/JobController.php
@@ -66,6 +66,12 @@ class JobController extends Controller
 
     //Update job data
     public function update(Request $request, Job $job) {
+
+        // Make sure logged in user is owner
+        if($job->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'], 
@@ -89,9 +95,18 @@ class JobController extends Controller
 
     //Delete Job
     public function destroy(Job $job) {
+        // Make sure logged in user is owner
+        if($job->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         //take job and call the delete method 
         $job->delete();
         return redirect('/')->with('message', 'Job deleted successfully');
 
+    }
+
+    //Manage Jobs
+    public function manage() {
+        return view('jobs.manage', ['jobs' => auth()->user()->jobs()->get()]); // auth()->user() gives the logged in user
     }
 }
